+++ common/src/main/java/immersive_melodies/client/animation/ItemAnimators.java
@@ -28,5 +28,6 @@ public class ItemAnimators {
         register(Common.locate("triangle"), new TriangleAnimator());
         register(Common.locate("trumpet"), new TrumpetAnimator());
         register(Common.locate("tiny_drum"), new TinyDrumAnimator());
+        register(Common.locate("vielle"), new VielleAnimator());
     }
 }